Update the `list` method to return all packages unless specified otherwise
This change brings back the behavior that pkgdb-cli had for pkgdb1 where
`list --user` return all the packages the user has ACLs for whatever the ACL.

This commit also adds a `--poc` argument allowing restricting the list of
packages returned to only the packages the user is point of contact for
(which was the default until now).

This allows listing the packages of a pseudo-user which we currently could
not do.

Relates to https://fedorahosted.org/pkgdb2/ticket/14#comment:12