Rework the get_bug method to ask for login only if needed
Before calling bugzilla we do not know if the user can log in or not. We
could call .login() but that would require asking the user for his/her
username and password everytime.
So instead we query bugzilla once, and based on the output we determine
if we need to call .login() or not (as bugzilla stores a token on disk
credentials are not necessary for every calls).