Adjust pkgdb-cli to allow specifying which pkgdb instance to query
This commit allows specifying which pkgdb, FAS and bugzilla instances to query
when running. This way one can run/test pkgdb-cli against a local pkgdb or FAS
instance.

In addition this commit introduce the --insecure argument allowing to ignore
checking the validity of the SSL certificate when running the queries.