Fix setting the co-maintainers when auto-approving new branches
This code path is hit when a package is requested to be created in
multiple Fedora branches. The package gets created in the first request
(most likely master) then the other branches are processed, but since the
package is in pkgdb, the action is re-qualified into request.branch and
that doesn't/didn't include adding the co-maintainers.

Fixes https://github.com/fedora-infra/pkgdb2/issues/385